// src/app/@core/model/user.model.ts
import { BaseModel } from './base.model';
import { Rolegroup } from './rolegroup.model';

export class User extends BaseModel{
    username?: string;
    password?: string;
    token?: string;
    fullname?: string;
    address?: string;
    phone?: string;
    email?: string;
    level?: ELevel;
    avatar?: string;
    birthday?: Date | number;
    gender?: number;
    status?: number;
    count?: number;
    roles?: Set<string> | Array<string>;
    classes?: Set<string> | Array<string>;
    roleGroups?:Set<string> | Array<string>;

    constructor() {
      super();
      this.username = '';
      this.password = '';
      this.phone = '';
      this.gender = 0;
      this.fullname = '';
    }
}

export enum ELevel{
  ANONYMOUS = 'ANONYMOUS',
  TEACHER = 'TEACHER',
  STUDENT = 'STUDENT',
  DIRECTOR = 'DIRECTOR',
  ADMINISTRATOR = 'ADMINISTRATOR'
}

export enum ERole {
  ROLE_USER_VIEW = 'ROLE_USER_VIEW',
  ROLE_USER_CREATE = 'ROLE_USER_CREATE',
  ROLE_USER_UPDATE = 'ROLE_USER_UPDATE',
  ROLE_USER_DELETE = 'ROLE_USER_DELETE',

  ROLE_BLOCK_VIEW = 'ROLE_BLOCK_VIEW',
  ROLE_BLOCK_CREATE = 'ROLE_BLOCK_CREATE',
  ROLE_BLOCK_UPDATE = 'ROLE_BLOCK_UPDATE',
  ROLE_BLOCK_DELETE = 'ROLE_BLOCK_DELETE',

  ROLE_CLASS_VIEW = 'ROLE_CLASS_VIEW',
  ROLE_CLASS_CREATE = 'ROLE_CLASS_CREATE',
  ROLE_CLASS_UPDATE = 'ROLE_CLASS_UPDATE',
  ROLE_CLASS_DELETE = 'ROLE_CLASS_DELETE',

  ROLE_MAJOR_VIEW = 'ROLE_MAJOR_VIEW',
  ROLE_MAJOR_CREATE = 'ROLE_MAJOR_CREATE',
  ROLE_MAJOR_UPDATE = 'ROLE_MAJOR_UPDATE',
  ROLE_MAJOR_DELETE = 'ROLE_MAJOR_DELETE',

  ROLE_MATHDESIGN_VIEW = 'ROLE_MATHDESIGN_VIEW',
  ROLE_MATHDESIGN_CREATE = 'ROLE_MATHDESIGN_CREATE',
  ROLE_MATHDESIGN_UPDATE = 'ROLE_MATHDESIGN_UPDATE',
  ROLE_MATHDESIGN_DELETE = 'ROLE_MATHDESIGN_DELETE',

  ROLE_CHAPTER_VIEW = 'ROLE_MATHDESIGN_VIEW',
  ROLE_CHAPTER_CREATE = 'ROLE_MATHDESIGN_CREATE',
  ROLE_CHAPTER_UPDATE = 'ROLE_MATHDESIGN_UPDATE',
  ROLE_CHAPTER_DELETE = 'ROLE_MATHDESIGN_DELETE',


  ROLE_TEST_VIEW = 'ROLE_TEST_VIEW',
  ROLE_TEST_CREATE = 'ROLE_TEST_CREATE',
  ROLE_TEST_UPDATE = 'ROLE_TEST_UPDATE',
  ROLE_TEST_DELETE = 'ROLE_TEST_DELETE',

  ROLE_SCORE_VIEW = 'ROLE_SCORE_VIEW',
  ROLE_SCORE_CREATE = 'ROLE_SCORE_CREATE',
  ROLE_SCORE_UPDATE = 'ROLE_SCORE_UPDATE',
  ROLE_SCORE_DELETE = 'ROLE_SCORE_DELETE',

  ROLE_KNOWLEGDE_VIEW = 'ROLE_KNOWLEGDE_VIEW',
  ROLE_KNOWLEGDE_CREATE = 'ROLE_KNOWLEGDE_CREATE',
  ROLE_KNOWLEGDE_UPDATE = 'ROLE_KNOWLEGDE_UPDATE',
  ROLE_KNOWLEGDE_DELETE = 'ROLE_KNOWLEGDE_DELETE',

  ROLE_STUDENT_VIEW = 'ROLE_STUDENT_VIEW',
  ROLE_STUDENT_CREATE = 'ROLE_STUDENT_CREATE',
  ROLE_STUDENT_UPDATE = 'ROLE_STUDENT_UPDATE',
  ROLE_STUDENT_DELETE = 'ROLE_STUDENT_DELETE',

  ROLE_TEACHER_VIEW = 'ROLE_TEACHER_VIEW',
  ROLE_TEACHER_CREATE = 'ROLE_TEACHER_CREATE',
  ROLE_TEACHER_UPDATE = 'ROLE_TEACHER_UPDATE',
  ROLE_TEACHER_DELETE = 'ROLE_TEACHER_DELETE',

  ROLE_ROLEGROUP_VIEW = 'ROLE_ROLEGROUP_VIEW',
  ROLE_ROLEGROUP_CREATE = 'ROLE_ROLEGROUP_CREATE',
  ROLE_ROLEGROUP_UPDATE = 'ROLE_ROLEGROUP_UPDATE',
  ROLE_ROLEGROUP_DELETE = 'ROLE_ROLEGROUP_DELETE',

  ROLE_TEST_TYPE_VIEW = 'ROLE_TEST_TYPE_VIEW',
  ROLE_TEST_TYPE_CREATE = 'ROLE_TEST_TYPE_CREATE',
  ROLE_TEST_TYPE_UPDATE = 'ROLE_TEST_TYPE_UPDATE',
  ROLE_TEST_TYPE_DELETE = 'ROLE_TEST_TYPE_DELETE',

  ROLE_POINT_VIEW = 'ROLE_POINT_VIEW',
  ROLE_POINT_CREATE = 'ROLE_POINT_CREATE',
  ROLE_POINT_UPDATE = 'ROLE_TPOINT_UPDATE',
  ROLE_POINT_DELETE = 'ROLE_POINT_DELETE',
}